Added to War and Blackjack
There is an issue being able to login --- only a guest account is reachable?? idk.

+++ src/BlackJackBoard.java
@@ -11,7 +11,7 @@ public class BlackJackBoard {
 	CardStack playerHand;
 	CardStack dealerHand;
 	CardStack deck;
-	int playerBet;
+	double playerBet;
 	
 	
 	public BlackJackBoard() {
@@ -39,11 +39,11 @@ public class BlackJackBoard {
 		return deck;
 	}
 	
-	public int getPlayerBet() {
+	public double getPlayerBet() {
 		return playerBet;
 	}
 
-	public void setPlayerBet(int playerBet) {
+	public void setPlayerBet(double playerBet) {
 		this.playerBet = playerBet;
 	}
